// types/playground.ts
export interface Model {
  name: string;
  model: string;
  provider: string;
}

export interface Agent {
  agent_id: string;
  name: string;
  description: string;
  model: Model;
  storage?: boolean;
}
// ...
export interface ComboboxAgent {
  value: string;
  label: string;
  model: {
    provider: string;
  };
  storage?: boolean;
}
// ...
export interface SessionEntry {
  session_id: string;
  title: string;
  created_at: number;
}